+++ client/src/pages/servapps/containers/newService.jsx
@@ -21,7 +21,6 @@ import { Link } from 'react-router-dom';
 import { smartDockerLogConcat, tryParseProgressLog } from '../../../utils/docker';
 import { LoadingButton } from '@mui/lab';
 import LogLine from '../../../components/logLine';
-import yaml from 'js-yaml';
 
 const preStyle = {
   backgroundColor: '#000',
@@ -85,7 +84,7 @@ const NewDockerService = ({service, refresh}) => {
     setLog([
       'Creating Service...                              ',
     ])
-    API.docker.createService(yaml.dump(service), (newlog) => {
+    API.docker.createService(service, (newlog) => {
       setLog((old) => smartDockerLogConcat(old, newlog));
       preRef.current.scrollTop = preRef.current.scrollHeight;
       if (newlog.includes('[OPERATION SUCCEEDED]')) {
@@ -119,7 +118,7 @@ const NewDockerService = ({service, refresh}) => {
       <pre style={preStyle} ref={preRef}>
         {!log.length && `
 # You are about to create the following service(s):
-${yaml.dump(service)}`
+${JSON.stringify(service, false ,2)}`
         }
         {log.map((l) => {
           return <LogLine message={tryParseProgressLog(l)} docker isMobile={!screenMin} />